fix: eliminacion de migracion pequeñas y creacion de una gran migracion
