--- alembic/versions/02a2c8959901_entidad_persona_grupo_retorno.py
@@ -1,40 +1,0 @@
-"""Entidad Persona_grupo_retorno
-
-Revision ID: 02a2c8959901
-Revises: 1c61b63e302b
-Create Date: 2026-04-29 22:36:31.686483
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '02a2c8959901'
-down_revision: Union[str, Sequence[str], None] = '1c61b63e302b'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('persona_grupo_retorno',
-    sa.Column('pgr_codigo', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('pe_codigo', sa.Integer(), nullable=False),
-    sa.Column('gr_codigo', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
-    sa.ForeignKeyConstraint(['pe_codigo'], ['persona.pe_codigo'], ),
-    sa.PrimaryKeyConstraint('pgr_codigo'),
-    sa.UniqueConstraint('pe_codigo', 'gr_codigo', name='uk1_uk2_persona_grupo_retorno')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('persona_grupo_retorno')
-    # ### end Alembic commands ###
