--- alembic/versions/1c61b63e302b_añadir_campo_reggr_anotacion.py
@@ -1,38 +1,0 @@
-"""añadir campo reggr_anotacion
-
-Revision ID: 1c61b63e302b
-Revises: dcdc7603e8d1
-Create Date: 2026-04-26 01:56:25.766848
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '1c61b63e302b'
-down_revision: Union[str, Sequence[str], None] = 'dcdc7603e8d1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('persona', sa.Column('pe_tipo_doc', sa.Enum('CC', 'CE', name='tipodoc'), nullable=False))
-    op.add_column('persona', sa.Column('pe_documento', sa.String(), nullable=False))
-    op.create_unique_constraint(None, 'persona', ['pe_documento'])
-    op.add_column('registro_retorno_grupo', sa.Column('reggr_anotacion', sa.String(length=500), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('registro_retorno_grupo', 'reggr_anotacion')
-    op.drop_constraint(None, 'persona', type_='unique')
-    op.drop_column('persona', 'pe_documento')
-    op.drop_column('persona', 'pe_tipo_doc')
-    # ### end Alembic commands ###
