--- alembic/versions/2072f7a2cd67_añadir_campo_anotación_a_registro_.py
@@ -1,32 +1,0 @@
-"""Añadir campo anotación a registro retorno individual
-
-Revision ID: 2072f7a2cd67
-Revises: 02a2c8959901
-Create Date: 2026-05-04 15:07:05.741889
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2072f7a2cd67'
-down_revision: Union[str, Sequence[str], None] = '02a2c8959901'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('registro_retorno', sa.Column('reg_anotacion', sa.String(length=500), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('registro_retorno', 'reg_anotacion')
-    # ### end Alembic commands ###
