--- alembic/versions/2c781b443c39_crear_tablas_de_grupos_de_retorno.py
@@ -1,92 +1,0 @@
-"""crear tablas de grupos de retorno
-
-Revision ID: 2c781b443c39
-Revises: efee06555c1c
-Create Date: 2026-04-25 02:10:05.626037
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2c781b443c39'
-down_revision: Union[str, Sequence[str], None] = 'efee06555c1c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('persona',
-    sa.Column('pe_codigo', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('pe_nombre', sa.String(length=100), nullable=False),
-    sa.Column('pe_apellido', sa.String(length=100), nullable=False),
-    sa.Column('pe_correo', sa.String(length=100), nullable=True),
-    sa.Column('pe_fecha_nacimiento', sa.String(length=100), nullable=False),
-    sa.PrimaryKeyConstraint('pe_codigo'),
-    sa.UniqueConstraint('pe_correo')
-    )
-    op.create_table('grupo_retorno',
-    sa.Column('gr_codigo', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('us_codigo_lider', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['us_codigo_lider'], ['usuario.us_codigo'], ),
-    sa.PrimaryKeyConstraint('gr_codigo')
-    )
-    op.create_table('registro_retorno',
-    sa.Column('reg_codigo', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('us_codigo', sa.Integer(), nullable=False),
-    sa.Column('re_codigo', sa.Integer(), nullable=False),
-    sa.Column('reg_num_hospedaje', sa.Integer(), nullable=False),
-    sa.Column('reg_num_transporte', sa.Integer(), nullable=False),
-    sa.Column('reg_num_parqueadero', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['re_codigo'], ['retorno.codigo'], ),
-    sa.ForeignKeyConstraint(['us_codigo'], ['usuario.us_codigo'], ),
-    sa.PrimaryKeyConstraint('reg_codigo'),
-    sa.UniqueConstraint('us_codigo', 're_codigo', name='uk1_uk2_registro_retorno')
-    )
-    op.create_table('registro_retorno_grupo',
-    sa.Column('regg_codigo', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('gr_codigo', sa.Integer(), nullable=False),
-    sa.Column('re_codigo', sa.Integer(), nullable=False),
-    sa.Column('reggr_num_hospedaje', sa.Integer(), nullable=False),
-    sa.Column('reggr_num_transporte', sa.Integer(), nullable=False),
-    sa.Column('reggr_num_parqueadero', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
-    sa.ForeignKeyConstraint(['re_codigo'], ['retorno.codigo'], ),
-    sa.PrimaryKeyConstraint('regg_codigo')
-    )
-    op.create_table('solicitud_grupo_retorno',
-    sa.Column('solgr_codigo', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('us_codigo', sa.Integer(), nullable=False),
-    sa.Column('gr_codigo', sa.Integer(), nullable=False),
-    sa.Column('pa_fecha_creacion', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
-    sa.ForeignKeyConstraint(['us_codigo'], ['usuario.us_codigo'], ),
-    sa.PrimaryKeyConstraint('solgr_codigo')
-    )
-    op.create_table('usuario_grupo_retorno',
-    sa.Column('ugr_codigo', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('us_codigo', sa.Integer(), nullable=False),
-    sa.Column('gr_codigo', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
-    sa.ForeignKeyConstraint(['us_codigo'], ['usuario.us_codigo'], ),
-    sa.PrimaryKeyConstraint('ugr_codigo'),
-    sa.UniqueConstraint('us_codigo', 'gr_codigo', name='uk1_uk2_usuario_grupo_retorno')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('usuario_grupo_retorno')
-    op.drop_table('solicitud_grupo_retorno')
-    op.drop_table('registro_retorno')
-    op.drop_table('registro_retorno_grupo')
-    op.drop_table('grupo_retorno')
-    op.drop_table('persona')
-    # ### end Alembic commands ###
