--- alembic/versions/6211ec8da47a_numero_parqueadero_por_vehiculo.py
@@ -1,42 +1,0 @@
-"""numero parqueadero por vehiculo
-
-Revision ID: 6211ec8da47a
-Revises: 2072f7a2cd67
-Create Date: 2026-05-05 13:05:49.200196
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '6211ec8da47a'
-down_revision: Union[str, Sequence[str], None] = '2072f7a2cd67'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('registro_retorno', sa.Column('reg_num_parqueadero_carro', sa.Integer(), nullable=False))
-    op.add_column('registro_retorno', sa.Column('reg_num_parqueadero_moto', sa.Integer(), nullable=False))
-    op.drop_column('registro_retorno', 'reg_num_parqueadero')
-    op.add_column('registro_retorno_grupo', sa.Column('reggr_num_parqueadero_carro', sa.Integer(), nullable=False))
-    op.add_column('registro_retorno_grupo', sa.Column('reggr_num_parqueadero_moto', sa.Integer(), nullable=False))
-    op.drop_column('registro_retorno_grupo', 'reggr_num_parqueadero')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('registro_retorno_grupo', sa.Column('reggr_num_parqueadero', sa.INTEGER(), autoincrement=False, nullable=False))
-    op.drop_column('registro_retorno_grupo', 'reggr_num_parqueadero_moto')
-    op.drop_column('registro_retorno_grupo', 'reggr_num_parqueadero_carro')
-    op.add_column('registro_retorno', sa.Column('reg_num_parqueadero', sa.INTEGER(), autoincrement=False, nullable=False))
-    op.drop_column('registro_retorno', 'reg_num_parqueadero_moto')
-    op.drop_column('registro_retorno', 'reg_num_parqueadero_carro')
-    # ### end Alembic commands ###
