+++ alembic/versions/90e12dc4b0ee_tablas_sprint_1_y_2.py
@@ -1,0 +1,189 @@
+"""tablas sprint 1 y 2
+
+Revision ID: 90e12dc4b0ee
+Revises: 
+Create Date: 2026-05-11 02:02:11.991095
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '90e12dc4b0ee'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('colonia',
+    sa.Column('co_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('co_pais', sa.String(), nullable=False),
+    sa.Column('co_departamento', sa.String(), nullable=True),
+    sa.Column('co_ciudad', sa.String(), nullable=True),
+    sa.Column('co_estado', sa.Enum('ACTIVA', 'INACTIVA', name='coloniaestado'), nullable=False),
+    sa.Column('lider_id', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('co_codigo')
+    )
+    op.create_table('persona',
+    sa.Column('pe_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('pe_tipo_doc', sa.Enum('CC', 'CE', name='tipodoc'), nullable=False),
+    sa.Column('pe_documento', sa.String(), nullable=False),
+    sa.Column('pe_nombre', sa.String(length=100), nullable=False),
+    sa.Column('pe_apellido', sa.String(length=100), nullable=False),
+    sa.Column('pe_correo', sa.String(length=100), nullable=True),
+    sa.Column('pe_fecha_nacimiento', sa.Date(), nullable=False),
+    sa.Column('pe_genero', sa.Enum('F', 'M', 'OTRO', name='genero'), nullable=False),
+    sa.PrimaryKeyConstraint('pe_codigo'),
+    sa.UniqueConstraint('pe_correo'),
+    sa.UniqueConstraint('pe_documento')
+    )
+    op.create_table('retorno',
+    sa.Column('codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('fecha_creacion', sa.DateTime(), nullable=False),
+    sa.Column('anio', sa.Integer(), nullable=False),
+    sa.Column('estado', sa.Enum('ACTIVO', 'EN_CURSO', 'FINALIZADO', name='retornoestado'), nullable=False),
+    sa.PrimaryKeyConstraint('codigo'),
+    sa.UniqueConstraint('anio', name='uk2_retorno')
+    )
+    op.create_index(op.f('ix_retorno_codigo'), 'retorno', ['codigo'], unique=False)
+    op.create_table('rol',
+    sa.Column('ro_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('ro_nombre', sa.String(), nullable=False),
+    sa.PrimaryKeyConstraint('ro_codigo'),
+    sa.UniqueConstraint('ro_nombre')
+    )
+    op.create_table('usuario',
+    sa.Column('us_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('us_fecha_creacion', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.Column('us_tipo_doc', sa.Enum('CC', 'CE', name='tipodoc'), nullable=False),
+    sa.Column('us_documento', sa.String(), nullable=False),
+    sa.Column('us_celular', sa.String(), nullable=False),
+    sa.Column('us_correo', sa.String(), nullable=False),
+    sa.Column('us_contrasenia', sa.String(), nullable=False),
+    sa.Column('co_codigo', sa.Integer(), nullable=True),
+    sa.Column('ro_codigo', sa.Integer(), nullable=False),
+    sa.Column('us_nombre', sa.String(), nullable=False),
+    sa.Column('us_apellido', sa.String(), nullable=False),
+    sa.Column('us_genero', sa.Enum('F', 'M', 'OTRO', name='genero'), nullable=False),
+    sa.Column('us_fecha_nacimiento', sa.Date(), nullable=False),
+    sa.Column('us_pais', sa.String(), nullable=False),
+    sa.Column('us_departamento', sa.String(), nullable=True),
+    sa.Column('us_ciudad', sa.String(), nullable=True),
+    sa.ForeignKeyConstraint(['co_codigo'], ['colonia.co_codigo'], ),
+    sa.ForeignKeyConstraint(['ro_codigo'], ['rol.ro_codigo'], ),
+    sa.PrimaryKeyConstraint('us_codigo'),
+    sa.UniqueConstraint('us_celular'),
+    sa.UniqueConstraint('us_correo'),
+    sa.UniqueConstraint('us_documento')
+    )
+    op.create_table('grupo_retorno',
+    sa.Column('gr_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('us_codigo_lider', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['us_codigo_lider'], ['usuario.us_codigo'], ),
+    sa.PrimaryKeyConstraint('gr_codigo')
+    )
+    op.create_table('parentesco',
+    sa.Column('pa_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('pa_fecha_creacion', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.Column('pa_estado', sa.Enum('pendiente', 'aceptada', 'rechazada', 'expirada', name='estadosolicitudparentesco'), nullable=False),
+    sa.Column('us_codigo_solicitante', sa.Integer(), nullable=False),
+    sa.Column('us_codigo_destinatario', sa.Integer(), nullable=False),
+    sa.Column('pa_tipo_parentesco', sa.Enum('padre', 'madre', 'hermano', 'hijo', 'abuelo', 'tio', 'primo', 'madrastra', 'padrastro', 'hijastro', 'conyuge', name='tipoparentesco'), nullable=False),
+    sa.ForeignKeyConstraint(['us_codigo_destinatario'], ['usuario.us_codigo'], ),
+    sa.ForeignKeyConstraint(['us_codigo_solicitante'], ['usuario.us_codigo'], ),
+    sa.PrimaryKeyConstraint('pa_codigo')
+    )
+    op.create_table('registro_retorno',
+    sa.Column('reg_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('us_codigo', sa.Integer(), nullable=False),
+    sa.Column('re_codigo', sa.Integer(), nullable=False),
+    sa.Column('reg_num_hospedaje', sa.Integer(), nullable=False),
+    sa.Column('reg_num_transporte', sa.Integer(), nullable=False),
+    sa.Column('reg_num_parqueadero_carro', sa.Integer(), nullable=False),
+    sa.Column('reg_num_parqueadero_moto', sa.Integer(), nullable=False),
+    sa.Column('reg_anotacion', sa.String(length=500), nullable=True),
+    sa.ForeignKeyConstraint(['re_codigo'], ['retorno.codigo'], ),
+    sa.ForeignKeyConstraint(['us_codigo'], ['usuario.us_codigo'], ),
+    sa.PrimaryKeyConstraint('reg_codigo'),
+    sa.UniqueConstraint('us_codigo', 're_codigo', name='uk1_uk2_registro_retorno')
+    )
+    op.create_table('solicitud_colonia',
+    sa.Column('so_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('us_codigo', sa.Integer(), nullable=False),
+    sa.Column('co_codigo', sa.Integer(), nullable=False),
+    sa.Column('so_estado', sa.Enum('pendiente', 'aceptada', 'rechazada', 'expirada', name='estadosolicitud'), nullable=False),
+    sa.Column('so_fecha_creacion', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['co_codigo'], ['colonia.co_codigo'], ),
+    sa.ForeignKeyConstraint(['us_codigo'], ['usuario.us_codigo'], ),
+    sa.PrimaryKeyConstraint('so_codigo')
+    )
+    op.create_index(op.f('ix_solicitud_colonia_so_codigo'), 'solicitud_colonia', ['so_codigo'], unique=False)
+    op.create_table('persona_grupo_retorno',
+    sa.Column('pgr_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('pe_codigo', sa.Integer(), nullable=False),
+    sa.Column('gr_codigo', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
+    sa.ForeignKeyConstraint(['pe_codigo'], ['persona.pe_codigo'], ),
+    sa.PrimaryKeyConstraint('pgr_codigo'),
+    sa.UniqueConstraint('pe_codigo', 'gr_codigo', name='uk1_uk2_persona_grupo_retorno')
+    )
+    op.create_table('registro_retorno_grupo',
+    sa.Column('regg_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('gr_codigo', sa.Integer(), nullable=False),
+    sa.Column('re_codigo', sa.Integer(), nullable=False),
+    sa.Column('reggr_num_hospedaje', sa.Integer(), nullable=False),
+    sa.Column('reggr_num_transporte', sa.Integer(), nullable=False),
+    sa.Column('reggr_num_parqueadero_carro', sa.Integer(), nullable=False),
+    sa.Column('reggr_num_parqueadero_moto', sa.Integer(), nullable=False),
+    sa.Column('reggr_anotacion', sa.String(length=500), nullable=True),
+    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
+    sa.ForeignKeyConstraint(['re_codigo'], ['retorno.codigo'], ),
+    sa.PrimaryKeyConstraint('regg_codigo')
+    )
+    op.create_table('solicitud_grupo_retorno',
+    sa.Column('solgr_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('us_codigo', sa.Integer(), nullable=False),
+    sa.Column('gr_codigo', sa.Integer(), nullable=False),
+    sa.Column('solgr_time_stamp', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.Column('solgr_estado', sa.Enum('PENDIENTE', 'ACEPTADO', 'RECHAZADO', 'EXPIRADO', name='solicitudgruporetornoestado'), nullable=False),
+    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
+    sa.ForeignKeyConstraint(['us_codigo'], ['usuario.us_codigo'], ),
+    sa.PrimaryKeyConstraint('solgr_codigo')
+    )
+    op.create_table('usuario_grupo_retorno',
+    sa.Column('ugr_codigo', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('us_codigo', sa.Integer(), nullable=False),
+    sa.Column('gr_codigo', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['gr_codigo'], ['grupo_retorno.gr_codigo'], ),
+    sa.ForeignKeyConstraint(['us_codigo'], ['usuario.us_codigo'], ),
+    sa.PrimaryKeyConstraint('ugr_codigo'),
+    sa.UniqueConstraint('us_codigo', 'gr_codigo', name='uk1_uk2_usuario_grupo_retorno')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('usuario_grupo_retorno')
+    op.drop_table('solicitud_grupo_retorno')
+    op.drop_table('registro_retorno_grupo')
+    op.drop_table('persona_grupo_retorno')
+    op.drop_index(op.f('ix_solicitud_colonia_so_codigo'), table_name='solicitud_colonia')
+    op.drop_table('solicitud_colonia')
+    op.drop_table('registro_retorno')
+    op.drop_table('parentesco')
+    op.drop_table('grupo_retorno')
+    op.drop_table('usuario')
+    op.drop_table('rol')
+    op.drop_index(op.f('ix_retorno_codigo'), table_name='retorno')
+    op.drop_table('retorno')
+    op.drop_table('persona')
+    op.drop_table('colonia')
+    # ### end Alembic commands ###
