--- alembic/versions/adebc859019a_agrega_columna_estado_a_tabla_colonia.py
@@ -1,45 +1,0 @@
-"""Agrega columna Estado a tabla Colonia
-
-Revision ID: adebc859019a
-Revises: 2072f7a2cd67
-Create Date: 2026-04-18 16:52:53.587340
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'adebc859019a'
-down_revision: Union[str, Sequence[str], None] = '2072f7a2cd67'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    # 1. Crear el tipo ENUM primero
-    coloniaestado_enum = sa.Enum('ACTIVA', 'INACTIVA', name='coloniaestado')
-    coloniaestado_enum.create(op.get_bind(), checkfirst=True)
-    
-    # 2. Agregar la columna con el enum
-    op.add_column('colonia', sa.Column('co_estado', coloniaestado_enum, nullable=False, server_default='ACTIVA'))
-    op.add_column('colonia', sa.Column('lider_id', sa.Integer(), nullable=True))
-    op.drop_column('colonia', 'lider')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('colonia', sa.Column('lider', sa.INTEGER(), autoincrement=False, nullable=True))
-    op.drop_column('colonia', 'lider_id')
-    op.drop_column('colonia', 'co_estado')
-    
-    # Eliminar el tipo ENUM
-    coloniaestado_enum = sa.Enum('ACTIVA', 'INACTIVA', name='coloniaestado')
-    coloniaestado_enum.drop(op.get_bind(), checkfirst=True)
-    # ### end Alembic commands ###
