--- alembic/versions/cf14f7b3d9c2_persona_genero_y_fecha_date.py
@@ -1,41 +1,0 @@
-"""persona genero y fecha date
-
-Revision ID: cf14f7b3d9c2
-Revises: 2072f7a2cd67
-Create Date: 2026-05-08 08:18:21.778402
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'cf14f7b3d9c2'
-down_revision: Union[str, Sequence[str], None] = '6211ec8da47a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('persona', sa.Column('pe_genero', sa.Enum('F', 'M', 'OTRO', name='genero'), nullable=False))
-    op.alter_column('persona', 'pe_fecha_nacimiento',
-               existing_type=sa.VARCHAR(length=100),
-               type_=sa.Date(),
-               existing_nullable=False,
-               postgresql_using="pe_fecha_nacimiento::date")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('persona', 'pe_fecha_nacimiento',
-               existing_type=sa.Date(),
-               type_=sa.VARCHAR(length=100),
-               existing_nullable=False)
-    op.drop_column('persona', 'pe_genero')
-    # ### end Alembic commands ###
