--- alembic/versions/dcdc7603e8d1_añadir_estados_a_solicitud_grupo_retorno.py
@@ -1,43 +1,0 @@
-"""Añadir estados a solicitud grupo retorno
-
-Revision ID: dcdc7603e8d1
-Revises: 2c781b443c39
-Create Date: 2026-04-25 04:03:30.346574
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'dcdc7603e8d1'
-down_revision: Union[str, Sequence[str], None] = '2c781b443c39'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    # Crear manualmente el tipo ENUM en PostgreSQL antes de añadir la columna
-    solicitud_estado_enum = postgresql.ENUM('PENDIENTE', 'ACEPTADO', 'RECHAZADO', 'EXPIRADO', name='solicitudgruporetornoestado')
-    solicitud_estado_enum.create(op.get_bind())
-
-    op.add_column('solicitud_grupo_retorno', sa.Column('solgr_time_stamp', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False))
-    op.add_column('solicitud_grupo_retorno', sa.Column('solgr_estado', sa.Enum('PENDIENTE', 'ACEPTADO', 'RECHAZADO', 'EXPIRADO', name='solicitudgruporetornoestado'), nullable=False))
-    op.drop_column('solicitud_grupo_retorno', 'pa_fecha_creacion')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('solicitud_grupo_retorno', sa.Column('pa_fecha_creacion', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=False))
-    op.drop_column('solicitud_grupo_retorno', 'solgr_estado')
-    op.drop_column('solicitud_grupo_retorno', 'solgr_time_stamp')
-
-    # Eliminar el tipo ENUM manualmente al hacer downgrade
-    postgresql.ENUM(name='solicitudgruporetornoestado').drop(op.get_bind())
-    # ### end Alembic commands ###
